+++ src/splat.ts
@@ -17,7 +17,7 @@ import {
 import { Element, ElementType } from "./element";
 import { Serializer } from "./serializer";
 import { State } from './splat-state';
-import { vertexShader, fragmentShader } from './shaders/splat-shader';
+import { vertexShader, fragmentShader } from './shaders/splat-outline-shader';
 import { TransformPalette } from './transform-palette';
 
 const vec = new Vec3();
@@ -390,6 +390,10 @@ class Splat extends Element {
             this._visible = value;
             this.scene.events.fire('splat.visibility', this);
         }
+    }
+
+    renderOutline() {
+        
     }
 }
